WIP: finally able to load initial values using redux-form initialValues along with own implementation of account-reducer.ts
