+++ src/reducers/index.ts
@@ -2,9 +2,11 @@ import { combineReducers } from 'redux';
 import { firebaseReducer } from 'react-redux-firebase';
 import { firestoreReducer } from 'redux-firestore';
 import { reducer as formReducer } from 'redux-form';
+import accountReducer from './account-reducer';
 
 export const rootReducer = combineReducers({
     firebase: firebaseReducer,
     firestore: firestoreReducer,
-    form: formReducer
+    form: formReducer,
+    account: accountReducer
 });
